+++ src/components/InfographicEditor/Editors/WaffleEditor.js
@@ -207,8 +207,9 @@ class WaffleEditor extends React.Component
                     cols={5}
                     onlyPositive={true}
                     onChange={(d, i) => { this._SetChartSettings('dynamicResize', 'width', d); }} 
+                    isDisabled={!this.props.cSettings.dynamicResize.isChecked}
                 />
-                <LabeledNumericTextField 
+                {false && <LabeledNumericTextField 
                     label='Height'
                     index='c-height'
                     initialValue={resize.height}
@@ -216,7 +217,7 @@ class WaffleEditor extends React.Component
                     cols={5}
                     onlyPositive={true}
                     onChange={(d, i) => { this._SetChartSettings('dynamicResize', 'height', d); }} 
-                />
+                />}
                 <LabeledNumericTextField 
                     label='Max icons per row: '
                     index='max'