+++ src/components/Notfications/NotificationEnums.js
@@ -1,0 +1,8 @@
+const Constants = {
+    NOTIFY: 'notify',
+    INFO: 'info',
+    ERROR: 'error',
+    SUCCESS: 'success',
+};
+
+export { Constants };